+++ src/schick/seg001.cpp
@@ -35,12 +35,12 @@
 #endif
 
 char cd_dummy0 = 2;
-unsigned short g_cd_init_successful = 0; /* {0,1} */
+signed int g_cd_init_successful = 0; /* {0,1} */
 static unsigned char cd_dummy1[4] = {0};
-static unsigned short g_cd_audio_repeat = 0; /* {0,1} */
+static signed int g_cd_audio_repeat = 0; /* {0,1} */
 static unsigned char cd_dummy2[4] = {0};
-static unsigned short g_cd_audio_paused = 0; /* {0,1} */
-static unsigned short g_cd_audio_track_bak = 0x7fff;
+static signed int g_cd_audio_paused = 0; /* {0,1} */
+static signed int g_cd_audio_track_bak = 0x7fff;
 static const int g_cd_audio_tracklist[15] = {
 	0x7fff,	0x7fff,	0x7fff, 0x7fff,
 	ARCHIVE_FILE_SUMMARY_XMI,
@@ -57,7 +57,7 @@ static const int g_cd_audio_tracklist[15] = {
 };
 uint32_t g_cd_skipmagic = 0L;
 
-static unsigned short g_cd_check_err_countdown = 5;
+static signed int g_cd_check_err_countdown = 5;
 static const char g_str_insert_cd[142] = "BITTE LEGEN SIE DIE \xf2SCHICKSALSKLINGE-CD\xf0 IN LAUFWERK %c: EIN. DIESE WIRD BEN\x99TIGT, DA DATEN W\x8eHREND DES SPIELS VON CD GELADEN WERDEN M\x9aSSEN.";
 static const char g_str_repeat[12] = "WIEDERHOLEN";
 static const char g_str_quit[8] = "BEENDEN";
@@ -66,13 +66,13 @@ static char g_str_cd_missing[204] = "DIESES PROGRAMM IST OHNE CD-ROM-LAUFWERK NI
 static const char g_str_cd_init[37] = "INITIALISIERE CD-ROM IN LAUFWERK %c:";
 static unsigned char g_unkn_004 = 0x00;
 
-static unsigned long g_cd_audio_pause_pos; // ds:0xbc38
-static unsigned long g_cd_audio_pause_tod; // ds:0xbc3c
-static unsigned short g_cd_audio_track; // ds:0xbc40
+static uint32_t g_cd_audio_pause_pos; // ds:0xbc38
+static uint32_t g_cd_audio_pause_tod; // ds:0xbc3c
+static signed int g_cd_audio_track; // ds:0xbc40
 static unsigned char g_unkn_071[8]; // ds:0xbc42
-static unsigned long g_cd_audio_tod; // ds:0xbc4a
-static unsigned long g_cd_audio_pos; // ds:0xbc4e
-static unsigned short g_cd_drive_num; // ds:0xbc52
+static uint32_t g_cd_audio_tod; // ds:0xbc4a
+static uint32_t g_cd_audio_pos; // ds:0xbc4e
+static signed int g_cd_drive_num; // ds:0xbc52
 static unsigned char g_unkn_072[4]; // ds:0xbc54
 
 /* static prototypes */
@@ -89,7 +89,7 @@ namespace M302de {
  * \return  {0 = no drive, 1 = at least one drive}
  */
 /* Borlandified and identical */
-static unsigned short CD_has_drives(void)
+static int16_t short CD_has_drives(void)
 {
 	/* al ==  0: return number of drive letters */
 	asm {
@@ -107,7 +107,7 @@ has_cd:
 }
 
 /* Borlandified and identical */
-static unsigned short CD_count_drives(void)
+static int16_t CD_count_drives(void)
 {
 	asm {
 		mov ax, 0x1500
@@ -119,7 +119,7 @@ static unsigned short CD_count_drives(void)
 }
 
 /* Borlandified and identical */
-static unsigned short CD_get_first_drive(void)
+static int16_t CD_get_first_drive(void)
 {
 	asm {
 		mov ax, 0x1500
@@ -131,7 +131,7 @@ static unsigned short CD_get_first_drive(void)
 }
 
 /* Borlandified and identical */
-static unsigned short CD_set_drive_num(void)
+static int16_t CD_set_drive_num(void)
 {
 	if (!CD_has_drives()) return 0;
 
@@ -186,7 +186,7 @@ leave_tod:
 }
 
 /* Borlandified and nearly identical */
-static void CD_audio_track_start(signed short track_num)
+static void CD_audio_track_start(signed int track_num)
 {
 	int32_t track_start;
 	int32_t track_end;
@@ -200,18 +200,18 @@ static void CD_audio_track_start(signed short track_num)
 				(cd_buf1[0x0e + track_num * 8]));
 
 		/* calculate track_start */
-		track_start = (60L * (unsigned long)cd_buf1[0x10 + track_num * 8]
+		track_start = (60L * (uint32_t)cd_buf1[0x10 + track_num * 8]
 			+ cd_buf1[0x0f + track_num * 8]) * 75L
 			+ cd_buf1[0x0e + track_num * 8];
 
 		/* calculate track_end */
 		if (cd_buf1[0x326] == track_num) {
 
-			track_end = (((unsigned long)cd_buf1[0x329] * 60
+			track_end = (((uint32_t)cd_buf1[0x329] * 60
 					+ cd_buf1[0x328]) * 75
 					+ cd_buf1[0x327]);
 		} else {
-			track_end = (((unsigned long)cd_buf1[0x18 + track_num * 8] * 60
+			track_end = (((uint32_t)cd_buf1[0x18 + track_num * 8] * 60
 					+ cd_buf1[0x17 + track_num * 8]) * 75
 					+ cd_buf1[0x16 + track_num * 8]);
 
@@ -219,8 +219,8 @@ static void CD_audio_track_start(signed short track_num)
 
 		track_start = track_end - track_start;
 
-		req[5].dummy4 = ((unsigned short)track_start) - 150;
-		req[5].dummy6 = (signed long)(track_start >> 16);
+		req[5].dummy4 = ((uint16_t)track_start) - 150;
+		req[5].dummy6 = (int32_t)(track_start >> 16);
 
 		CD_driver_request(&req[5]);
 
@@ -250,7 +250,7 @@ void CD_enable_repeat(void)
 }
 
 /* Borlandified and identical */
-int CD_bioskey(const int cmd)
+signed int CD_bioskey(const signed int cmd)
 {
 #if defined(__BORLANDC__)
 	CD_enable_repeat();
@@ -336,7 +336,7 @@ void CD_audio_play(void)
 /* Borlandified and identical */
 static void CD_0432(void)
 {
-	signed short track_num;
+	signed int track_num;
 
 	if (!g_cd_init_successful) return;
 
@@ -345,7 +345,7 @@ static void CD_0432(void)
 	cd_buf1[0x324] = 0x0a;
 	CD_driver_request(&req[2]);
 
-	track_num = (unsigned short)cd_buf1[0x325];
+	track_num = (uint16_t)cd_buf1[0x325];
 
 	while (cd_buf1[0x326] >= track_num)
 	{
@@ -361,14 +361,14 @@ static void CD_0432(void)
 #endif
 
 struct dummy15 {
-	signed short a[15];
+	int16_t a[15];
 };
 
 /* Borlandified and identical */
-void CD_set_track(const int index)
+void CD_set_track(const signed int index)
 {
 #if defined(__BORLANDC__)
-	signed short i;
+	signed int i;
 	struct dummy15 tracks = *(struct dummy15*)g_cd_audio_tracklist;
 
 	for (i = 0; i < 15; i++)
@@ -400,11 +400,11 @@ void CD_set_track(const int index)
 
 #if defined(__BORLANDC__)
 /* Borlandified and identical */
-static signed short CD_read_exe(char *path)
+static int16_t CD_read_exe(char *path)
 {
-	int handle;
-	signed short buffer;
-	unsigned short nread;
+	signed int handle;
+	signed int buffer;
+	uint16_t nread;
 
 	/* skip read check */
 	if (g_cd_skipmagic == 0x682772e4L) return 1;
@@ -425,7 +425,7 @@ static signed short CD_read_exe(char *path)
 /* Borlandified and identical */
 static void CD_insert_msg(void)
 {
-	signed short answer;
+	signed int answer;
 	char str[160];
 
 	sprintf(str, g_str_insert_cd, g_cd_drive_num + 'A');
@@ -445,7 +445,7 @@ static void CD_insert_msg(void)
 }
 
 /* Borlandified and identical */
-static signed short CD_harderr_handler(void)
+static signed int CD_harderr_handler(void)
 {
 	if (g_cd_check_err_countdown == 0)
 	{
@@ -483,7 +483,7 @@ void CD_check(void)
 }
 
 /* Borlandified and identical */
-int CD_init(void)
+signed int CD_init(void)
 {
 #if defined(__BORLANDC__)
 	char str[80];
